requirements.txt
# LangGraph and LangChain Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# Google Gemini Integration
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
google-api-core>=2.19.0

# Model Context Protocol
mcp>=1.0.0

# Retry Logic
tenacity>=8.2.0

# Persistence
aiosqlite>=0.20.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
